// src/app/core/models/user.model.ts
export interface User {
    id: number;
    name: string;
    username: string;
    email: string;
    address: Address;
    phone: string;
    website: string;
    company: Company;
}

export interface Geolocation {
    lat: string;
    lng: string;
}
  
export interface Address {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo: Geolocation;
}
  
export interface Company {
    name: string;
    catchPhrase: string;
    bs: string;
}
  